Python/CRUD/web-app/api/requirements.txt
bcrypt==4.2.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cors==1.0.1
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
future==1.0.0
gevent==24.11.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
psycopg2==2.9.10
PyJWT==2.10.1
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
requests-file==2.1.0
tldextract==5.1.3
urllib3==2.3.0
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
